requirements-dev.txt
pylint==3.3.9
pylint-django==2.6.1
mypy==1.1.1
djangorestframework-stubs==1.10.0
types-Pillow==10.2.0.20240822
types-reportlab==4.2.0.20240525
types-qrcode==7.4.0.20240408
tblib==3.1.0
